pyproject.toml
[tool.poetry]
name = "Test"
version = "0.1.0"
description = ""
authors = ["Victor Elceaninov <victor.knkproject@gmail.com>"]
license = "GNU GENERAL PUBLIC LICENSE Version 3"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
uvicorn = {extras = ["standart"], version = "^0.14.0"}
pymongo = "^3.11.4"
python-dotenv = "^0.17.1"
motor = "^2.4.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
dnspython = "^2.1.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
